+++ gulpfile.ts
@@ -1,0 +1,2 @@
+export * from './packages/cli/migrate-make-styles'
+export * from './packages/cli/build/typescript'
