+++ packages/cli/migrate-make-styles/index.ts
@@ -1,0 +1,3 @@
+export async function count() {}
+count.displayName = 'ms-count'
+count.description = 'Count how many makeStyles needs to be migrated'
