+++ packages/cli/package.json
@@ -18,7 +18,6 @@
     "@types/inquirer": "^7.3.3",
     "@types/lodash": "^4.14.171",
     "@types/proper-lockfile": "^4.1.2",
-    "gulp": "^4.0.2",
     "ts-node": "^10.1.0"
   }
 }
