+++ packages/cli/tsconfig.json
@@ -7,5 +7,5 @@
     },
     "transpileOnly": true
   },
-  "include": ["build", "locale-kit"]
+  "include": ["build", "locale-kit", "utils", "migrate-make-styles"]
 }
